Fix incomplete moving ExportFormat from main to serialize
