+++ src/main.rs
@@ -1,13 +1,12 @@
 //! Entry point of the program.
 use nickel::error::{Error, IOError, SerializationError};
 use nickel::program::Program;
-use nickel::serialize;
+use nickel::{serialize, serialize::ExportFormat};
 use nickel::term::{RichTerm, Term};
 use nickel::{repl, repl::rustyline_frontend};
 use std::io::Write;
 use std::path::PathBuf;
-use std::str::FromStr;
-use std::{fmt, fs, io, process};
+use std::{fs, io, process};
 // use std::ffi::OsStr;
 use structopt::StructOpt;
 
@@ -21,46 +20,6 @@ struct Opt {
     file: Option<PathBuf>,
     #[structopt(subcommand)]
     command: Option<Command>,
-}
-
-impl std::default::Default for ExportFormat {
-    fn default() -> Self {
-        ExportFormat::Json
-    }
-}
-
-impl fmt::Display for ExportFormat {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            Self::Raw => write!(f, "raw"),
-            Self::Json => write!(f, "json"),
-            Self::Yaml => write!(f, "yaml"),
-            Self::Toml => write!(f, "toml"),
-        }
-    }
-}
-
-#[derive(Clone, Eq, PartialEq, Debug)]
-pub struct ParseFormatError(String);
-
-impl fmt::Display for ParseFormatError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "unsupported export format {}", self.0)
-    }
-}
-
-impl FromStr for ExportFormat {
-    type Err = ParseFormatError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_ref() {
-            "raw" => Ok(ExportFormat::Raw),
-            "json" => Ok(ExportFormat::Json),
-            "yaml" => Ok(ExportFormat::Yaml),
-            "toml" => Ok(ExportFormat::Toml),
-            _ => Err(ParseFormatError(String::from(s))),
-        }
-    }
 }
 
 /// Available subcommands.
